refactor: move checkGameOver and checkGameWin to minesweeperUtils
